+++ src/components/CreateStudyComponents/CreateBackground.js
@@ -1,14 +1,14 @@
-import styles from "./CreateBackground.module.css";
-import React, { useState } from "react";
-import backgroundGreen from "../img/background_1.png";
-import backgroundYe from "../img/background_2.png";
-import backgroundBlu from "../img/background_3.png";
-import backgroundPink from "../img/background_4.png";
-import backgroundTable from "../img/background_5.png";
-import backgroundSun from "../img/background_6.png";
-import backgroundRain from "../img/background_7.png";
-import backgroundPlan from "../img/background_8.png";
-import backgroundIcon from "../img/background_icon.png";
+import styles from './CreateBackground.module.css';
+import React, { useState } from 'react';
+import backgroundGreen from '../../img/background_1.png';
+import backgroundYe from '../../img/background_2.png';
+import backgroundBlu from '../../img/background_3.png';
+import backgroundPink from '../../img/background_4.png';
+import backgroundTable from '../../img/background_5.png';
+import backgroundSun from '../../img/background_6.png';
+import backgroundRain from '../../img/background_7.png';
+import backgroundPlan from '../../img/background_8.png';
+import backgroundIcon from '../../img/background_icon.png';
 
 const images = [
   backgroundGreen,
@@ -28,7 +28,7 @@ function Background({ handleChange }) {
     setSelectedIndex(index);
 
     const value = images[index];
-    const name = "background";
+    const name = 'background';
     handleChange({ name, value });
   };
 
@@ -47,7 +47,7 @@ function Background({ handleChange }) {
             {index === selectedIndex && (
               <img
                 src={backgroundIcon}
-                alt="이미지선택아이콘"
+                alt='이미지선택아이콘'
                 className={styles.iconImage}
               />
             )}
