+++ src/components/CreateStudyComponents/CreateFooter.module.css
@@ -1,4 +1,4 @@
-@import "../css/reset.css";
+@import "../../css/reset.css";
 
 .footer {
   position: relative;
