+++ src/components/CreateStudyComponents/CreateForm.module.css
@@ -1,4 +1,4 @@
-@import '../css/reset.css';
+@import '../../css/reset.css';
 
 .background {
   margin: 40px auto;
