+++ src/components/CreateStudyComponents/CreateInput.module.css
@@ -1,4 +1,4 @@
-@import '../css/reset.css';
+@import '../../css/reset.css';
 
 .text {
   font-size: 18px;
